requirements.txt
google-cloud-bigquery==3.35.1
google-cloud-storage==3.2.0
jafgen==0.4.14
pandas==2.3.1

